backend/scripts/ocr/requirements.txt
paddleocr==2.7.3
paddlepaddle==2.6.2
pdf2image==1.17.0
pillow==10.4.0
numpy==1.26.4
